+++ spec/unit/mongoid/default_scope_spec.rb
@@ -4,16 +4,56 @@ describe Mongoid::DefaultScope do
 
   describe ".default_scope" do
 
-    subject { Acolyte.all }
+    let(:criteria) do
+      Acolyte.all
+    end
 
-    its(:options) { should == {:sort => [[:name, :asc]]} }
-  end
+    it "applies the scope to any criteria" do
+      criteria.options.should eq({ :sort => [[ :name, :asc ]] })
+    end
 
-  context "combined with a named scope" do
+    context "when combining with a named scope" do
 
-    subject { Acolyte.active }
+      let(:scoped) do
+        Acolyte.active
+      end
 
-    its(:options) { should == {:sort => [[:name, :asc]]} }
-    its(:selector) { should == {:status => "active"} }
+      it "applies the default scope" do
+        scoped.options.should eq({ :sort => [[ :name, :asc ]] })
+      end
+    end
+
+    context "when calling unscoped" do
+
+      let(:unscoped) do
+        Acolyte.unscoped
+      end
+
+      it "does not contain the default scoping" do
+        unscoped.options.should eq({})
+      end
+
+      context "when applying a named scope after" do
+
+        let(:named) do
+          Acolyte.unscoped.active
+        end
+
+        it "does not contain the default scoping" do
+          named.options.should eq({})
+        end
+
+        context "when applying multiple scopes after" do
+
+          let(:multiple) do
+            named.named
+          end
+
+          it "does not contain the default scoping" do
+            multiple.options.should eq({})
+          end
+        end
+      end
+    end
   end
 end